Change project organization to io.zipkin.brave.play
