+++ play-zipkin-tracing/core/src/main/scala/brave/play/TraceData.scala
@@ -1,0 +1,7 @@
+package brave.play
+
+/**
+ * Class for a trace data take over between requests.
+ * @param span the present span
+ */
+case class TraceData(span: brave.Span)
