+++ play-zipkin-tracing/core/src/main/scala/brave/play/ZipkinTraceConfig.scala
@@ -1,0 +1,8 @@
+package brave.play
+
+object ZipkinTraceConfig {
+  val AkkaName = "zipkin-trace-context"
+  val ServiceName = "trace.service-name"
+  val ZipkinBaseUrl = "trace.zipkin.base-url"
+  val ZipkinSampleRate = "trace.zipkin.sample-rate"
+}
