--- play-zipkin-tracing/core/src/main/scala/zipkin/play/TraceData.scala
@@ -1,8 +1,0 @@
-package zipkin.play
-
-/**
- * Class for a trace data take over between requests.
- * @param span the present span
- */
-case class TraceData(span: brave.Span)
-
