--- play-zipkin-tracing/core/src/main/scala/zipkin/play/ZipkinTraceConfig.scala
@@ -1,8 +1,0 @@
-package zipkin.play
-
-object ZipkinTraceConfig {
-  val AkkaName = "zipkin-trace-context"
-  val ServiceName = "trace.service-name"
-  val ZipkinBaseUrl = "trace.zipkin.base-url"
-  val ZipkinSampleRate = "trace.zipkin.sample-rate"
-}
