+++ play-zipkin-tracing/core/src/test/scala/brave/play/MapSetterTest.scala
@@ -1,0 +1,18 @@
+package brave.play
+
+import brave.propagation.Propagation
+import brave.test.propagation.PropagationSetterTest
+
+/** Tests [[ZipkinTraceServiceLike.mapSetter]] is a valid setter */
+class MapSetterTest extends PropagationSetterTest[collection.mutable.Map[String, String], String] {
+
+  import collection.JavaConverters._
+
+  override val keyFactory = Propagation.KeyFactory.STRING
+  override val carrier = collection.mutable.Map[String, String]()
+
+  override def setter = ZipkinTraceServiceLike.mapSetter
+
+  override def read(carrier: collection.mutable.Map[String, String], key: String) =
+    carrier.get(key).toList.asJava
+}
