--- play-zipkin-tracing/play/src/main/scala/zipkin/play/implicits/ZipkinTraceImplicits.scala
@@ -1,37 +1,0 @@
-package zipkin.play.implicits
-
-import brave.Span
-import play.api.mvc.RequestHeader
-import zipkin.play.actor.ActorTraceSupport.ActorTraceData
-import zipkin.play.{TraceData, ZipkinTraceServiceLike}
-
-trait ZipkinTraceImplicits {
-
-  // for injection
-  val tracer: ZipkinTraceServiceLike
-
-  /**
-   * Creates a trace data including a span from request headers.
-   *
-   * @param req the HTTP request header
-   * @return the trace data
-   */
-  implicit def request2trace(implicit req: RequestHeader): TraceData = {
-    TraceData(
-      span = tracer.toSpan(req.headers)((headers, key) => headers.get(key))
-    )
-  }
-
-  /**
-   * Creates a trace data including a span from request headers for Akka actor.
-   *
-   * @param req the HTTP request header
-   * @return the trace data
-   */
-  implicit def request2actorTrace(implicit req: RequestHeader): ActorTraceData = {
-    val span = tracer.toSpan(req.headers)((headers, key) => headers.get(key))
-    val oneWaySpan = tracer.newSpan(Some(span.context())).kind(Span.Kind.CLIENT)
-    ActorTraceData(span = oneWaySpan)
-  }
-
-}
